Attached is a patch to autoconf which makes the following changes:
*) Adds support for wxWindows 2.5 (--enable-wx25)
*) Adds support for release version of pgadmin3 (--enable-release)

*DO NOT* use the --enable-release flag yet to configure.  I still have
to work out what a "release" build really means, and what a "debug"
build means.  It may require the use of the "-static" flag to gcc
again, which causes a lot of dependency problems which will take some
time to track down.  --enable-wx25 is safe and test, however (at least
by me!).

ahp


git-svn-id: svn://svn.pgadmin.org/trunk@1259 a7884b65-44f6-0310-8a51-81a127f17b15