I wanted to build a .app bundle for pgadmin, but I found no support for this in the build system. I added an option --build-bundle that basically overwrites --prefix to $(pwd)tmp, and added a PgAdminIII.app target to Makefile.am. The built bundle includes the pgadmin3 binary (named PgAdminIII, all shared libraries that binary depends on (system libs in /usr/lib and /System are excluded), the documentation and the ui/ folder. This code is in pgadmin3-buildbundle.diff
I added code to pgAdmin3.cpp to find the ui and doc folder inside the
bundle, using the SystemPathsCF class from wx (only when __WXMAC__, of
course). Using just SystemPaths would be cleaner, as it should default
to sensible locations on all plattforms, but as of wxmac 2.5.4,
SystemPaths refers to the unix-implemenation, not to the osx
implemenation. This code is in pgadmin3-usebundle.diff
[Florian G. Pflug]


git-svn-id: svn://svn.pgadmin.org/trunk@4030 a7884b65-44f6-0310-8a51-81a127f17b15