Here's two more minor fixes to the edit data grid.
First is related to the copy changes I recently made. Since I modified 
the code to copy the cell the cursor is on if there is no highlight, 
this change makes the Copy button in the toolbar enabled at all times.

The other is related to editing text cells. It looks like someone 
half-implemented a check to only update the database if the contents of 
the cell actually changed. This fix finishes that work. This 
functionality was already in place for numeric cells.
[Edward Di Geronimo Jr.]

git-svn-id: svn://svn.pgadmin.org/trunk@4995 a7884b65-44f6-0310-8a51-81a127f17b15