Andreas reports: > When triggering autocomplete on spaces (e.g. empty query > window), a memory damage will be triggered. If ignored, the > full keyword listbox will popup. > Happens with VC6, not gtk.
Apologies for the delay in dealing with this, been way too busy.
Attached is a fix. A buffer in _complete_from_list() wasn't allocaed
with enough space for the terminating NULL. I've fixed a similar issue
in complete_create_command - I wasn't able to trigger it, but anyway.

It really should hapen on gtk as well - most likely just luck that it
didn't corrupt anything needed. [Magnus Hagander]


git-svn-id: svn://svn.pgadmin.org/trunk@5234 a7884b65-44f6-0310-8a51-81a127f17b15