Fix minor bug reported by Erwin:
# One more cornercase for "Find and Replace". Try to replace multiple occurrences of  'abc' with 'abcd' and see for yourself ..
Seems like "Replace" continues the search at the _beginning_ of the newly inserted text. If the search string is included in the replace string, we end up in an endless loop.
Might be easy to fix by continuing the search _after_ the replaced string. ("Find" seems to do it right.)


git-svn-id: svn://svn.pgadmin.org/trunk@5639 a7884b65-44f6-0310-8a51-81a127f17b15