build a Debugger.app within the pgAdmin3.app resources directory in the Mac App bundle.
This does introduce one downside:

The existing complete-bundle.sh code works fine, and rewrites the paths
to the wx libraries as expected, however the debugger crashes when run.
As best I can figure, this is because although the debugger has the
correct relative paths to the libraries, some of those (wx) libraries
are interdependent, and when run by the debugger executable, the
relative paths between them are incorrect.

To resolve this until a better solution is found, I've modified the
build system such that static linking is forced when building an app
bundle. This also has the advantage of reducing the bundle size a little  :-) 




git-svn-id: svn://svn.pgadmin.org/trunk@6164 a7884b65-44f6-0310-8a51-81a127f17b15