Remove bundled copies of PostgreSQL/Slony docs to make it easier to localise the rest. Instead allow the use of  local or online copies of the PostgreSQL, EnterpriseDB or Slony docs, each independently of the other. Per discussion on -hackers.
This implements the infrastructure to handle multiple documentation sources, and add code to use EnterpriseDB docs when connected to an EDB server. Configuration options are included, as well as validation/cleanup of user entered doc paths. Still to complete is the documentation for the configuration options, and the auto-discovery of local docs.

git-svn-id: svn://svn.pgadmin.org/trunk@6302 a7884b65-44f6-0310-8a51-81a127f17b15