Add support to the debugger for capture of OUT/INOUT params from edbspl functions. This requires that we use an EnterpriseDB libpq to work due to the protocol extensions included for callable statement support.
To ensure this is optional, on Windows we attempt to runtime link to the extra libpq 
functions, whereas on *nix we use a configure time check to see what libpq we are
building against. Standard libpq builds will work just fine, except they cannot
catch the out params from stored procs.

Whilst we're at it, clean up some of the autoconf code that was testing functionality
of libpq.so even if we were static linking against libpq.a


git-svn-id: svn://svn.pgadmin.org/trunk@6352 a7884b65-44f6-0310-8a51-81a127f17b15