Don't run a separate query for the database description for each database, instead fold it into the main query.
Also, don't refresh the default schema until actually connecting to the
database - previous code ran it in the maintenance database once for each
database in the system, and then again in the actual database after
connection.

Fixes a large performance regression on systems with a very large number
of databases.


git-svn-id: svn://svn.pgadmin.org/branches/REL-1_8_0_PATCHES@7476 a7884b65-44f6-0310-8a51-81a127f17b15