The change to using long options for pg_dump and friends seem to have broken it, at least on linux. But I assume it still worked on at least osx *or* windows, since that's where Dave would've tested it :-)
It would more or less never work, since it tried to put a double-quote
in the actual filename *before* the / specifying the root directory..

Here's a patch that removes the = sign after the parameter names. That
makes it work on linux again. I would appreciate it if somebody can test
this patch on windows and/or osx as well to confirm it doesn't break
things there.

[Magnus Hagander]

git-svn-id: svn://svn.pgadmin.org/trunk@7717 a7884b65-44f6-0310-8a51-81a127f17b15