Please find the patch for an issue related to the debugger in gtk (linux).
Please find the steps to reproduce this issue:
1. Right click any pl/pgsql function without any parameter. Select
   Debugging->Debug.
2. It will open the debugger window.
3. Select File->Exit from menu.
4. Press 'yes' for the below question:
   "Are you sure you wish to abort the debugging session?
    This will abort the function currently being debugged."
5. Close main window.
6. This will gives you the following message and wait for all threads to be
   getting killed.
   "TIME WITH TIMEZONE: Debug: 1 threads were not terminated by the application."

I found that - we're creating two dbgPgConn object, one in ctlCodeWindow and the
other in dlgDirectDbg. Both creates their own threads (dbgThread object).

When we abort debugging, somehow it is not calling the Close function of the
dlgDirectDbg only on linux (wxGTK). Hence, the thread created in dlgDirectDbg is
not get killed/stopped in this case.

[Ashesh Vashi]


git-svn-id: svn://svn.pgadmin.org/trunk@7838 a7884b65-44f6-0310-8a51-81a127f17b15