Default role support in server configs, quick connect, backups and restore.
Rolename parameter can be specified in each server's saved configuration.
When a rolename is specified issue a SET role=<rolename> command after a
successfull connection to the server. This allows using pgAdmin's gui to
create/alter/delete objects in accordance with a security policy that
forbids logging in directly as superuser, but instead specifies that you
should log into an admin account and then SET ROLE to the superuser.

It also makes it possible to specify the superuser account at backup
and restore dialogs (pg_dump and pg_restore version >= 8.4 needed).

Patch from Benedek László.