Don't send the Greenplum query for each user, but import it in the big roles query.
The CreateObjects() method launched a query to get all informations, but not
the Greenplum one. Then, another query is launched for every user if we are
connected to a Greenplum database. This patch merges the two query so that
only one query is launched.

Report from Little, Douglas. Patch from me. Checked by Chuck McDevitt.