Fix an issue with tab keypress on frmQuery.
On Windows, a user pressing tab on the scratch pad of the frmQuery window does
not insert a tab character on the scratch pad. It does insert it on the query
textbox.

This fix checks if the user presses tab in the query textbox or in the scratch
pad, and reacts accordingly. If he presses tab on another component, the event
is ignored.