Make the dlgSelectConnection combobox read/write
When a user launches the dlgSelectConnection dialog from the query tool, all
combobox are read-only, forcing the user to select one of the available items
on each combobox.

When a user launches pgAdmin with -q or -S command line options, he first gets
the dlgSelectConnection dialog. As no connection is already made, the only
thing the combox box will have depends on the previous quick connect action.
So, the combobox should be read/write, so that he can change the values.

It used to work in 1.10, but was broken since 1.12 because of one of my
patches.

Reported by Steffen Kuhn. Patch by him, reworked by me.