Add to the main window a link to each query window
frmQuery windows can be launched from differents parts of the code. Some
didn't take care to add the new frmQuery window to the frames of the main
window. This patch takes care of this in the launch code and in the frmStatus
window.

Fixes #223. Report by Franz Fritsche. Patch from Steffen Kuhn, with a minor
addition from me.