Get rid of the schema name and double quotes
The format_type stored procedure returns the formated name of the type
referenced by its OID. On PostgreSQL releases above 7.4, it includes the
schema name and double quotes if needed. We need to suppress the schema name
and the double quotes for a pgDatatype as we already have the schema name in a
specific variable (the "schema" one). If we don't do so, some mehods return
the type name with the schema name doubled.

Report from Valentine Gogichashvili.