Skip searching a R/W cell on a view
If a user hits Enter at the end of a row, the edit grid searches the next
read/write cell of the displayed object. Unfortunately, there is no read/write
cell when displaying a view. This patch checks first if the object's kind is a
table. If it is, it proceeds to the search of a writable cell. Otherwise, it
simply picks up the next cell.

Report from David Jarvis.