Fix for failed new connection in query tool
When a user tries to open a new connection in the query tool, and that the new
connection fails, the conn variable still has the invalid connection handle.
And this prevents us from getting back to the previous older one. So, we need
to use a temporary variable to store the new (and perhaps invalid) connection
handle.

Report from Erwin Brandstetter.