Fix the reverse-engineered query for triggers
The new "OF column" syntax is only available for UPDATE, and we should check
the args only if the number of args is not zero.

Set the bytea_output parameter to its old value on 9.0 right after a
connection. The tgargs column is of type bytea and bytea_output, new parameter
in 9.0, makes it quite unreadable if we don't set bytea_output to 'escape'.

Report from Christopher A Hotchkiss.

Fixed #240.