Fix connection color on the query tool
The color used the old connection informations. In fact, we call two times the
OnChangeConnection() method. First time to choose another connection and add
the new one to the combobox, and another time to really change the connection.
IOW, we change the colors despite the fact we actually didn't change the
connection. So we can't check the connection with the connection global
variable. We need to give the connection local variable to the
GetServerColour() method. So, this method gets a connection parameter.

Report from Valentine Gogichashvili.

Fixed #241.