Fix two deletes of the same object in a query tool
When a user executes a query and wants to save the results, he clicks on the
disk-with-a-green-arrow icon. If he clicks on the Cancel button, pgAdmin
crashes. The issue was that we deleted two times the qi->toFileExportForm
object. Once in the frmQuery code and once in the qi destroy method. Fixed by
deleting the one in the frmQuery code.

Report from Jean-Michel Souchard.

Fixes #255.