Use -w switch when available for dump/restore
If a user tries to dump a database within pgAdmin, and if there is no .pgpass
file, pgAdmin is frozen waiting for a password noone can enter. The -w switch
allows the use of .pgpass file but, if this file is missing and the
authentication method asks for a password, pg_dump/pg_restore will fail rather
than block pgAdmin.

Report from Anssi Kaariainen.

Implements #270.