Stop losing focus in Find dialog if nothing more
This patch fixes an issue with the focus. When on the Find box, at the end of a document, a message box is displayed telling the user that it reached the end. After the user presses Enter, focus is lost where as it should still be on the Find dialog.