Re-indent source per new rules
Mega-run of the new "make style" command, to enforce
coding style. This is likely to cause conflicts with most
pending patches, but there shouldn't be many - and they should
be easy to resolve.

From now on, it's recommended to run "make style" before a commit
if you do development on Unix, to keep the style up. We'll also
make a regular run before releases to cover things not handled
during development.