So, I changed the locale on my setup to "zh_CN.UTF-8" and tried to restore into a database using a backup from earlier. When I click on "Display objects" tab, I see an error being displayed, although the actual restore happens just fine. Note that no error occurs in English locales.
Took me a while to figure this one out. The issue is with frmRestore::OnEndProcess() function. It naively avoids lines containing ";" or "P" as the first letter. So in the English locale, the last line about the "Process returned exit code 0" is handled. However in other locales, this will be translated and hence the last line will be inspected leading to a spurious error.

The fix is to avoid inspecting the last line.