Fix segfault when connecting to a database
The issue appears when you first connect to a database, and then disconnect
from it. If you try to connect once again to it, you get a segmentation fault
because pgAdmin will try to get the database statistics informations from a
non-existent connection.

The patch adds a check on the connection status.

Report from Steffen Kuhn.