Fix a crash bug when trying to create a new table on 8.3
Instead of deleting the autovacuum notebook page, pgAdmin deleted the
constraints notebook page. So, when the user clicked OK, pgAdmin tried to find
changes in the constraints widgets which didn't exist anymore.

This bug was introduced by me when I added the new Definition tab to most of
the dialogs. Will do a complete check of "DeletePage" in each of the dialogs.