Current behaviour of the ExplainPopup, requires a mouse click on the ExplainShape to show it. A lot of user will not able to figure out that until they accidentally click on that shape or read the documents carefully.
Also - current implementation does not allow to use the ExplainCanvas in a modal dialog.
For some experiment, I tried to use the ExplainCanvas in a modal dialog and clicked on a shape.

A lot of time on linux, this scenario resulted in a crash or unstable pgadmin application.

The attached patch enables the ExplainPopup to behave like the tooltip for those shapes.
As per my conversation with Dave, I have extended the ExplainPopup using wxTransientPopup or, wxFrame
(where not available wxPopWindow i.e. Mac), just like the wxTipWindow, and also worked on the border
around the ExplainText.