Overhaul of type attributes modification
If you add, delete, rename or change the type, collation or precision
of a composite type attribute, it deletes all of them then adds them
all back in.  Obviously attribute additions, deletions and modifications
may only occur for types under PostgreSQL 9.1, but it's a bit extreme
to actually drop all types if you're adding a new one in, or just
removing one.  Also, if you rename a type's element or change the type
of a type's element, it drops all attributes again then re-adds them,
just to have an attribute with a different name.

So this patch tries to do it right, with the minimum required queries.