Fix weird behaviour when changing column's collation
If a user opens a table properties dialog, and tries to change the collation
of a column, he gets a DROP COLUMN followed by a ADD COLUMN statements.
If a user opens a column properties dialog (without the table properties
dialog first), and tries to change the collation, he gets nothing.

Both behaviour are wrong. We should get an ALTER TABLE... ALTER COLUMN...
TYPE... COLLATION... This patch fixes both issues.