Work-around the wxGTK bug with wxAuiNotebook in frmQuery
See http://forums.wxwidgets.org/viewtopic.php?f=23&t=31090 for details.

When a user tries to get back in the first tab of the frmQuery after modifying
a graphical query, pgAdmin seems frozen and non responsive. Most of the OS
seems the same. Only the keyboard is available and can get us back to a normal
behaviour.

The bug seems to be within wxGTK (as it only happens on Linux), so this patch
is only a work-around.

Initial report from Colin Beckingham.