Fix two bugs in default security panel
1. When a user clicks on an item in the list, there was no update on the combobox
and the checkboxes.

2. This one was reported by Kashif Zeeshan:
"On Privileges Tab Page when a role is added without Privileges (click on
Add/Change? button without selecting any check box in Privileges section) then
one row is added with role name in the list and when the same role is changed
by adding the privileges in the Privileges section then the same row is
updated but incase of Default Privileges Tab (Table, Sequence and Functions
sub tabs) when the same above steps are followed then two rows are added, one
with role name and second with role name and the assigned privileges. One Row
must be added on Default Privileges Tab same as on Privileges Tab."

Fixes #335