Fix bad query in the query tool
Issue reported in http://forums.enterprisedb.com/posts/list/2831.page:
On a quiet server with 12 sessions, all are <IDLE>, but some are in the color
for idle and some are in the color for slow query. It is my understanding that
Slow Queries is for queries running over 10 seconds, but these are idle and
should reflect that color.