Fix crash when opening a dlgType
At the opening of the dialog, we do a CheckChange for no reason (at least, I don't see one) and this triggers a crash because the connection variable isn't already set. We could set the connection before the CheckChange but I prefer to delete the not-needed CheckChange call.

Per "report" from Vladimir Kokovic.