Fix compilation on Debian
We used a wxWidgets internal macro (ATTRIBUTE_PRINTF_1), that Debian get rid
of. And wxWidgets 2.9 doesn't have this macro anymore.

Actually, it appears we don't need it. It works without the macro, so the patch just drops it.

Reports from Gerfried Fuchs, and Philippe Rimbault.
Many thanks to Dave Page, and Olly Bets, for their comments.
Fixes Debian Bug #646454.