Fix loss of column changes through table properties when re-editing
When the user validates its changes on the column properties dialog, we
create another pgColumn containing all the changes made. We store it in a new
column of the columns listview. When the user wants to see the changed column
properties dialog, we use this new object to make sure the column dialog shows
the changed, but not yet applied, properties.

Report from Thom Brown.
Fixes #330.