Support new/changed pg_stat_activity fields
With 4f42b546fd87a80be30c53a0f2c897acb826ad52, PostgreSQL has a new column in
pg_stat_activity. The state column shows the state of the connection. The
column current_query has been renamed to query. It shows the current query if
the state is active, and the previous query if the state is not active.

This patch adds support for all these changes. The state column is displayed.
The query is displayed too, even if it is the previous one. The query start
column is only displayed when the state is active (either because the state
column is set to active in 9.2, or because there is '<IDLE>' in the beginning
of the query field in previous releases.