Fix the autovacuum widgets in the table dialog
When a user selects another language for the UI, some of the values of the
autovacuum are set to 0 instead of their actual values because of the decimal
separator (dot) which is not valid for this language (happens with french and
polish for example).

This fix uses always the dot. To do that, I had to change the text widgets so
that they are not confined to numeric values.

Report from Bartek.