Fix the copy of a query when application_name is unset
The copy of a query tris to fetch the database name, so that when we launch
the query tool, we connect to the right database. Since 9.0, with the addition
of the application_name column, database is not anymore column 1, but column
2, when the user is connected to 9.0+ database. So, when we fetch the database
name, we need to look into the right column.