Make sure that queries use LIKE with _ escaped
For example:

SELECT nspname
FROM pg_namespace
WHERE nspname NOT LIKE 'pg_%'
  AND nspname != 'information_schema'
ORDER BY nspname

will get every schema whose name isn't information_schema and doesn't begin
with pg (the _ character is a joker character when used with LIKE, meaning
"replace-me-with-one-character-whatever-it-is").

So, we need to escape the underscore to get rid of its special meaning.