Fix the restore of a window position when using multiscreen
The previous code used to get the geometry of the good display and check
position and size of the dialog against the geometry of the display.
Unfortunately, the position and size of the dialog take into account the
global geometry. So we need to compare it with the global geometry, which this
patch does.

Per report from Steffen Decker.