Fix the SQL when adding a new index without a name
When a user adds an index without setting its name, but with a comment or the
CLUSTER option, the two specific SQL commands to set the comment and the
CLUSTER option don't have the name of the index (as we don't know it yet), and
thus fail.

If we don't have a name, we now disable both widgets (comment and cluster),
and we don't take care of their value in GetSql method().

Per report from Bartosz Dmytrak