Add support for setting db/user config on dlgRole
Previously, a user could only add/remove/update db/user configuration in the
database properties dialog, which was a weird choice I made when adding
support for db/user configuration. Now, a user can also do that from the role
properties dialog.

It also fixes an issue with older releases that a user can trigger when adding
an equal sign in a value of a parameter (for example a schema name containing
an equal sign).