Don't reset breakpoints when folding code
In the debugger window, when you click on a cross on the left margin to
fold or unfold a block of code, we also set or clear a breakpoint on
that line. That's clearly not the way it's supposed to work - should
only set/clear breakpoints when you click on the line number or the
margin that displays the red breakpoint symbol.