Fix the check status after closing a message box
The status has three possible values: YES, NO, and CANCEL. Most of the code
assume that the user clicks NO, and so the code checks for NO. But the user
can cancel the dialog, and as te code checks for NO, the code continues to
execute just as if the user clicked OK: table truncated, connection closed,
query killed without the explicit order of the user.

Patch from nikicat (https://github.com/nikicat), heavily changed by me.