Fix the FK covering index retrieval
This was broken since 373c6b949ba8e4cbe3d3756f27a6a7dc68e2869b commit. If the
column's type doesn't an opclass, it still gets an extra space in its
definition. When we look for the index, we check the column definition, and
this extra space makes the definitions look different.

Per a report from Nikolay Ulyanitsky.