Fix the search query
It tried to look into the pg_user_mapping catalog. A regular user cannot read
this table, hence a query error when searching objects while connected as a
regular user.

This patch adds a new IsSuperuser() method to the pgConn class. It calls this
method to check if it should add the UNION clause with the SQL/MED tables
(which pg_user_mapping is a part of).

Per a report from Chandrakant G.