Save new comments on already existing columns when editing a table
If I use the object browser context menu for an existing table and
choose "Properties...', I get a dialog that looks identical to dialog
for creating the table.  It allows me to add a new column to the table
just like in the table creation case, and it asks for a comment on the
new column.  But that comment gets thrown away.  The SQL tab reflects
the ALTER TABLE statement to add the column, but not the additional
statement to record the comment on that new column.

This patch fixes this. Unfortunately, it has a side-effect. When someone
changes a column property (but not the comment), and the column already had a
comment, it adds a COMMENT query with the old comment. So it's one useless
query. Not a big deal, and fixing this would be a huge work.

Per a report from Jeff Janes.