Improve the copy handling on the Edit Grid so that it works more consistently and predictably.
Previously, different copy functions may be called if using the
toolbar button vs. the keyboard shortcut, which led to inconsistent
behaviour. Further, when a cell was in Edit Mode, the entire contents
would be copied, where the user would normally only expect the
highlighted text (if any) to be copied.