If the function, which is being debugged, raise too many notices, pgadmin will become unresponsive.
As wxTextCtrl is not able to cop up with frequent AppendText calls, I have to introduce a timer to show the
current message(s) after the timer stops. That allows to avoid the frequent call to the AppendText function.
Appending those message to the current message instead of calling the function.