Don't use --as-needed on Solaris.
As /usr/bin/ld on illumos doesn't (yet) support --as-needed, this
patch to acinclude.m4 will suppress the LDFLAGS -Wl,-as-needed
passed when gcc is used as the compiler in the "bash bootstrap" command.

This patch assumes that all gcc variants running on illumos (gcc3 and
gcc4) are configured with --without-gnu-ld --with-ld=/usr/bin/ld,
which after brief research, seems typically the case.