Fix for changes in Apple XCode's tools output.
While building pgAdmin3 on OS X, we observed that the script
complete-bundle.sh, greps for the string "Mach-O i386 executable"
from the output of file command. This does not work on my OS X 10.8,
where the file shows the output in a different order.

$ file pgAdmin3
pgAdmin3: Mach-O executable i386

So, the grep fails there. Attached herewith is the patch that should
work for "Mach-O i386 executable" and as well as "Mach-O executable
i386" outputs.