Honour the PGPASSFILE environment variable (if provided)
Current implementation saves/retrieves password from the default .pgpass
(on non-windows)/pgpass.conf (on windows). But - actually when
connections are made libpq honours PGPASSFILE. Hence - we need to do the
same through pgAdmin.

While fetching the configuration file path, we will create the required
directory structure (if permissible and does not exist) to keep the
PGPASSFILE in it. And, create the file represented by PGPASSFILE with
0600 permission (if does not exist).

PATCH: Prasad Somwanshi
Reworked by: Sanket Mehta
Reviewed by: Ashesh Vashi & Sanet Mehta