Add classid filter to queries on pg_depend
There are a number of cases where queries in both pgadmin4 and pgadmin3
are done against pg_depend but falis to constrain the query on classid.

In particular, if for example a constraint and a sequence exists with
the same oid (which is perfectly valid, as they are in different
tables), a column will suddenly refer sequences that are completely
incorrect. When we look up sequences, we have to make sure we
only match dependencies against pg_class.

This is a pgadmin3 version of the same fix that was applied to pgadmin4.